% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{tableXMLs}
\alias{tableXMLs}
\title{Return xml links by organization}
\usage{
tableXMLs(filepath = NULL)
}
\description{
Returns a list of two tables: a Publishers table with details on each publisher, and an XMLs table with links to all available xmls with the associated name and publisher.
}
